# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Sistema de Facturacion Medica API
  version: 0.6.0-draft
  description: |
    Contrato API contract-first en construccion por lotes.
    Lote 1 implementado: auth + base global.
    Lote 2 implementado: patients, doctors, insurance.
    Lote 3 implementado: appointments + medical-records.
    Lote 4 implementado: invoices + payments.
    Lote 5 implementado: catalog.
servers:
  - url: http://localhost:8080
    description: Local dev
...
paths:
...
  /api/v1/patients:
    post:
      tags: [Patients]
      operationId: patients_create
      summary: Crear paciente
      description: Requiere rol ADMIN o RECEPTIONIST.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientCreateRequest'
      responses:
        '201':
          description: Paciente creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '400': { $ref: '#/components/responses/BadRequestError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '409': { $ref: '#/components/responses/ConflictError' }
        '422': { $ref: '#/components/responses/UnprocessableError' }
    get:
      tags: [Patients]
      operationId: patients_list
      summary: Listar pacientes
      parameters:
        - name: lastName
          in: query
          required: false
          schema: { type: string }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Lista paginada de pacientes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagePatientSummaryResponse'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
  /api/v1/patients/{id}:
    put:
      tags: [Patients]
      operationId: patients_update
      summary: Actualizar paciente
      description: Requiere rol ADMIN o RECEPTIONIST.
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientUpdateRequest'
      responses:
        '200':
          description: Paciente actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '400': { $ref: '#/components/responses/BadRequestError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '422': { $ref: '#/components/responses/UnprocessableError' }
    get:
      tags: [Patients]
      operationId: patients_get_by_id
      summary: Obtener paciente por id
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Paciente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /api/v1/patients/dni/{dni}:
    get:
      tags: [Patients]
      operationId: patients_get_by_dni
      summary: Obtener paciente por DNI
      parameters:
        - name: dni
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Paciente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /api/v1/patients/search:
    get:
      tags: [Patients]
      operationId: patients_autocomplete
      summary: Buscar pacientes por texto
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Lista de coincidencias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientSummaryResponse'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
  /api/v1/patients/{id}/appointments:
    get:
      tags: [Patients]
      operationId: patients_get_appointments
      summary: Listar citas del paciente
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Lista paginada de citas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageAppointmentSummaryResponse'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /api/v1/patients/{id}/policies:
    get:
      tags: [Patients]
      operationId: patients_get_policies
      summary: Listar polizas del paciente
      parameters:
        - $ref: '#/components/parameters/Id'
        - name: onlyActive
          in: query
          required: false
          schema: { type: boolean }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Lista paginada de polizas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageInsurancePolicySummaryResponse'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /api/v1/patients/{id}/invoices:
    get:
      tags: [Patients]
      operationId: patients_get_invoices
      summary: Listar facturas del paciente
      parameters:
        - $ref: '#/components/parameters/Id'
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [DRAFT, PENDING, PARTIAL_PAID, PAID, CANCELLED, OVERDUE]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Lista paginada de facturas del paciente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageInvoiceListViewResponse'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
...
  /api/v1/medical-records/patient/{patientId}:
    get:
      tags: [Medical Records]
      operationId: medical_records_list_by_patient
      summary: Listar expedientes por paciente
      parameters:
        - name: patientId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Lista paginada de expedientes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageMedicalRecordResponse'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        description: Formato de Spring Data, por ejemplo `lastName,asc`
  schemas:
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
...
    MedicalRecordResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        patientId: { type: string, format: uuid }
        patientFirstName: { type: string }
        patientLastName: { type: string }
        appointmentId: { type: string, format: uuid }
        vitalSigns:
          type: object
          additionalProperties: true
          nullable: true
        physicalExam: { type: string, nullable: true }
        clinicalNotes: { type: string }
        recordDate: { type: string, format: date-time }
        createdAt: { type: string, format: date-time, nullable: true }
        updatedAt: { type: string, format: date-time, nullable: true }
...
    InvoiceListViewResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        patientId: { type: string, format: uuid }
        patientFirstName: { type: string }
        patientLastName: { type: string }
        invoiceNumber: { type: string }
        total: { type: number }
        patientResponsibility: { type: number }
        status:
          type: string
          enum: [draft, pending, partial_paid, paid, cancelled, overdue]
        issueDate: { type: string, format: date }
        dueDate: { type: string, format: date }
        createdAt: { type: string, format: date-time, nullable: true }
...
    PatientCreateRequest:
      type: object
      required: [dni, firstName, lastName, birthDate, gender, phone]
      properties:
        dni:
          type: string
          maxLength: 20
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        birthDate:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other, prefer_not_to_say]
        phone:
          type: string
          maxLength: 20
        email:
          type: string
          format: email
          maxLength: 100
        address:
          type: string
          maxLength: 200
        bloodType:
          type: string
          maxLength: 5
        allergies:
          type: string
          maxLength: 500
    PatientUpdateRequest:
      type: object
      properties:
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        birthDate:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other, prefer_not_to_say]
        phone:
          type: string
          maxLength: 20
        email:
          type: string
          format: email
          maxLength: 100
        address:
          type: string
          maxLength: 200
        bloodType:
          type: string
          maxLength: 5
        allergies:
          type: string
          maxLength: 500
    PatientResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        dni: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        birthDate: { type: string, format: date }
        gender:
          type: string
          enum: [male, female, other, prefer_not_to_say]
        phone: { type: string }
        email: { type: string, format: email, nullable: true }
        address: { type: string, nullable: true }
        bloodType: { type: string, nullable: true }
        allergies: { type: string, nullable: true }
        createdAt: { type: string, format: date-time, nullable: true }
        updatedAt: { type: string, format: date-time, nullable: true }
    PatientSummaryResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        dni: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        phone: { type: string }
...
    InsurancePolicySummaryResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        policyNumber: { type: string }
        providerName: { type: string }
        coveragePercentage: { type: number }
        isActive: { type: boolean }
    AppointmentSummaryResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        patientFirstName: { type: string }
        patientLastName: { type: string }
        doctorFirstName: { type: string }
        doctorLastName: { type: string }
        scheduledAt: { type: string, format: date-time }
        status:
          type: string
          enum: [scheduled, confirmed, in_progress, completed, cancelled, no_show]
    PagePatientSummaryResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PatientSummaryResponse'
        totalElements: { type: integer }
        totalPages: { type: integer }
        size: { type: integer }
        number: { type: integer }
...
    PageInsurancePolicySummaryResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/InsurancePolicySummaryResponse'
        totalElements: { type: integer }
        totalPages: { type: integer }
        size: { type: integer }
        number: { type: integer }
    PageAppointmentSummaryResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AppointmentSummaryResponse'
        totalElements: { type: integer }
        totalPages: { type: integer }
        size: { type: integer }
        number: { type: integer }
    PageMedicalRecordResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/MedicalRecordResponse'
        totalElements: { type: integer }
        totalPages: { type: integer }
        size: { type: integer }
        number: { type: integer }
...
    PageInvoiceListViewResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceListViewResponse'
        totalElements: { type: integer }
        totalPages: { type: integer }
        size: { type: integer }
        number: { type: integer }
...
  responses:
    BadRequestError:
      description: Solicitud invalida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: No autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Sin permisos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Conflicto de datos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableError:
      description: Regla de negocio violada
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
security:
  - bearerAuth: []